Moved enums into their own files
