+++ src/NuGet.Services.Validation/NuGet.Services.Validation.csproj
@@ -41,9 +41,12 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Entities\Certificate.cs" />
+    <Compile Include="Entities\CertificateStatus.cs" />
     <Compile Include="Entities\CertificateValidation.cs" />
     <Compile Include="Entities\Package.cs" />
     <Compile Include="Entities\PackageSignature.cs" />
+    <Compile Include="Entities\PackageSignatureStatus.cs" />
+    <Compile Include="Entities\PackageSigningStatus.cs" />
     <Compile Include="Entities\PackageValidation.cs" />
     <Compile Include="Entities\PackageValidationSet.cs" />
     <Compile Include="Entities\ValidationContext.cs" />
